refactor(examples): splitting src of DonutChart example; split
